requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-filter==2.4.0
djangorestframework==3.12.2
factory-boy==3.1.0
Faker==4.18.0
idna==2.10
Markdown==3.3.3
mysqlclient==2.0.1
oauthlib==3.1.0
pipdeptree==1.0.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
